apiproject/requirements.txt
asgiref==3.5.2
autocrop==1.3.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cmake==3.24.0
cycler==0.11.0
Django==4.0.6
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.34.4
gunicorn==20.1.0
idna==3.3
imageio==2.21.1
install==1.3.5
kiwisolver==1.4.4
matplotlib==3.5.2
networkx==2.8.5
numpy==1.23.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
panda==0.3.1
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.0
six==1.16.0
sqlparse==0.4.2
tifffile==2022.8.3
urllib3==1.26.11
whitenoise==6.2.0
